# docs/reference/openapi.yaml
openapi: 3.0.3
info:
  title: Momentum Trading Assistant API
  version: 3.0.0
  description: |
    SUPPORTING REFERENCE FILE (MANUALLY MAINTAINED)

    This OpenAPI specification is a supporting reference representation of the API.
    Canonical API contract truth lives in:
      - docs/specs/api_contracts/

    If this file disagrees with the canonical Markdown API contracts,
    the Markdown contracts take precedence.

    Versioning:
    - This file's version matches the canonical API contract version documented in:
      docs/specs/api_contracts/README.md

    Key principles:
    - Decision-support tool with deterministic server-side logic
    - Backend is single source of truth for calculations
    - Exit prices are user-provided (broker executions)
    - Multi-currency support (GBP base; USD/GBP native values where applicable)
servers:
  - url: https://api.example.com
    description: Production server
  - url: http://localhost:8000
    description: Development server
...
paths:
  /portfolio:
    get:
      tags: [Portfolio]
      summary: Get portfolio overview
      description: |
        Returns complete portfolio overview with live prices, cash balance,
        and open positions.
      responses:
        "200":
          $ref: "#/components/responses/OkPortfolioOverview"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /portfolio/pre-entry-validation:
    get:
      tags: [Portfolio]
      summary: Advisory pre-entry rule validation (SI-01)
      description: >
        Non-blocking advisory check of 5 pre-entry rules: regime gate, cash constraint,
        sector concentration, earnings proximity, sizing validity.
        All checks are advisory — results do not block trade plan submission.
        §13 compliant: decision support only. strategy_rules.md §4.2.
        Decision record: docs/product/decisions/decisions--2026-05-19__release-v3.8--SI-01-section13-review.md
        Canonical contract: docs/specs/api_contracts/pre_entry_validation.md v1.0.0.
      parameters:
        - in: query
          name: ticker
          required: true
          schema:
            type: string
          description: Ticker symbol
        - in: query
          name: quantity
          required: true
          schema:
            type: number
          description: Proposed number of shares
        - in: query
          name: market
          required: false
          schema:
            type: string
            enum: [US, UK]
            default: US
          description: Market (US or UK)
        - in: query
          name: entry_price
          required: false
          schema:
            type: number
            nullable: true
          description: Entry price — enables sizing validity check
        - in: query
          name: stop_price
          required: false
          schema:
            type: number
            nullable: true
          description: Stop price — enables sizing validity check
      responses:
        "200":
          description: Advisory validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  data:
                    type: object
                    properties:
                      ticker:
                        type: string
                      market:
                        type: string
                      quantity:
                        type: number
                      advisory_status:
                        type: string
                        enum: [pass, warn, fail]
                      override_required:
                        type: boolean
                      checks:
                        type: array
                        items:
                          type: object
                          properties:
                            rule:
                              type: string
                            status:
                              type: string
                              enum: [pass, warn, fail, skipped]
                            detail:
                              type: string
                            severity:
                              type: string
        "422":
          description: Missing required query params
...
components:
  schemas:
    ApiOk:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          enum: [ok]
        data:
          description: Payload varies by endpoint.
          type: object
...
    ApiError:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
          enum: [error]
        message:
          type: string
    PortfolioOverview:
      type: object
      properties:
        cash: { type: number }
        cash_balance: { type: number }
        total_value: { type: number }
        open_positions_value: { type: number }
        total_pnl: { type: number }
        initial_value: { type: number }
        net_deposits: { type: number }
        live_fx_rate: { type: number }
        last_updated: { type: string, format: date-time }
        current_drawdown_percent:
          type: number
          format: float
          description: "Percentage decline from all-time peak portfolio value. Result is <= 0.0; zero means portfolio is at peak. Defaults to 0.0 when no portfolio_history exists."
        peak_portfolio_value:
          type: number
          format: float
          description: "All-time high total_value from portfolio_history (GBP, all-time, not period-scoped). Defaults to 0.0 when no portfolio_history exists."
        positions:
          type: array
          items: { $ref: "#/components/schemas/PositionSummary" }
    PositionSummary:
      type: object
      description: >
        Portfolio position summary object as returned by GET /portfolio.
        v1.9.0 field list per docs/specs/api_contracts/portfolio_endpoints.md.
        current_price is always GBP (USD converted via live_fx_rate for US positions).
        entry_price is in the instrument's native currency (USD for US, GBP for UK).
        pnl_pct is the percentage return relative to entry price (signed %).
        grace_days_remaining is max(0, 10 - holding_days) when grace_period is true; null otherwise.
      properties:
        id: { type: string, format: uuid }
        ticker: { type: string }
        market: { type: string, enum: [US, UK] }
        entry_date: { type: string, format: date }
        entry_price: { type: number }
        shares: { type: number }
        current_price: { type: number, description: "Always GBP regardless of instrument currency" }
        current_value: { type: number, description: "current_price × shares, in GBP" }
        pnl: { type: number }
        pnl_pct: { type: number, description: "Percentage return relative to entry price (signed %)" }
        current_stop: { type: number, description: "Current stop loss price in GBP" }
        holding_days: { type: integer }
        status: { type: string, enum: [open, closed] }
        display_status: { type: string, enum: [GRACE, PROFITABLE, LOSING] }
        fx_rate: { type: number, description: "FX rate stored at time of position entry" }
        grace_period: { type: boolean }
        grace_days_remaining: { type: integer, nullable: true }
        live_fx_rate: { type: number, description: "FX rate used for current price conversion" }
...
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication via X-API-Key header
  responses:
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    OkPortfolioOverview:
      description: Portfolio overview retrieved successfully
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiOk"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/PortfolioOverview"
...
security:
  - ApiKey: []